Issue #4261 (require --force to set unknown svn: prop names) Fix Windows build warnings and errors.
* subversion/svn_private_config.hw: Define HAVE_QSORT_R
* subversion/svn/props.c (svn_cl__check_svn_prop_name): Cast default score
   to unsigned int on assignment.


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@1412948 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/svn/props.c
@@ -356,7 +356,7 @@ svn_cl__check_svn_prop_name(const char *propname, svn_boolean_t revprop,
       propkeys[i] = &propbuf[i];
       propbuf[i].name.data = proplist[i];
       propbuf[i].name.len = strlen(proplist[i]);
-      propbuf[i].score = -1;
+      propbuf[i].score = (unsigned int)-1;
 #if !HAVE_QSORT_R
       propbuf[i].context = &context;
 #endif
